migrate miopen functionaility from worker_interface to fin_class, remove old functions
