+++ tuna/fin_class.py
@@ -31,18 +31,22 @@ import os
 import tempfile
 import functools
 import paramiko
+import random
+from time import sleep
 from sqlalchemy import func as sqlalchemy_func
 from sqlalchemy.exc import IntegrityError, InvalidRequestError  #pylint: disable=wrong-import-order
 
 from tuna.worker_interface import WorkerInterface
-from tuna.db_tables import connect_db
 from tuna.dbBase.sql_alchemy import DbSession
 from tuna.metadata import FIN_CACHE
 from tuna.metadata import INVERS_DIR_MAP
 from tuna.fin_utils import compose_config_obj
-from tuna.tables import DBTables
+from tuna.fin_utils import get_fin_slv_status
 from tuna.config_type import ConfigType
-from tuna.utils.db_utility import get_solver_ids, session_retry
+from tuna.utils.db_utility import session_retry
+from tuna.utils.db_utility import get_solver_ids, get_id_solvers
+
+MAX_JOB_RETRIES = 10
 
 
 class FinClass(WorkerInterface):
@@ -53,37 +57,25 @@ class FinClass(WorkerInterface):
   def __init__(self, **kwargs):
     """Constructor"""
     super().__init__(**kwargs)
-    allowed_keys = set([
-        'fin_steps', 'local_file', 'fin_outfile', 'fin_infile', 'machine',
-        'docker_name', 'version', 'config_type', 'label', 'session_id'
-    ])
+    allowed_keys = set(['fin_steps', 'local_file', 'fin_infile', 'fin_outfile'])
     self.__dict__.update((key, None) for key in allowed_keys)
 
-    connect_db()
-    self.all_configs = []
-    self.jobid_to_config = {}
-    self.fin_list = []
-    self.arch_list = []
     self.supported_fin_steps = ["get_solvers", "applicability"]
+    self.fin_steps = []
     _, self.local_file = tempfile.mkstemp()
     self.fin_infile = self.local_file.split("/tmp/", 1)[1] + ".json"
     _, self.local_output = tempfile.mkstemp()
     self.fin_outfile = self.local_output.split("/tmp/", 1)[1] + ".json"
-    self.fin_steps = []
-    self.machine = None
-    self.docker_name = None
-    self.cnx = None
-    self.config_type = ConfigType.convolution if self.config_type is None else self.config_type
-    self.label = None
-    self.session_id = None
 
+    self.solver_id_map = get_solver_ids()
+    _, self.id_solver_map = get_id_solvers(
+    )  #hyphenated names used by miopen::solver.ToString()
+    self.all_configs = []
+    self.fin_list = []
     self.multiproc = False
 
     self.__dict__.update(
         (key, value) for key, value in kwargs.items() if key in allowed_keys)
-
-    self.dbt = DBTables(session_id=self.session_id,
-                        config_type=self.config_type)
 
   def chk_abort_file(self):
     """Checking presence of abort file to terminate processes immediately"""
@@ -326,7 +318,6 @@ class FinClass(WorkerInterface):
 
   def insert_applicability(self, session, json_in):
     """write applicability to sql"""
-    _, solver_id_map_h = get_solver_ids()
     for elem in json_in:
       if "applicable_solvers" in elem.keys():
         #remove old applicability
@@ -342,7 +333,7 @@ class FinClass(WorkerInterface):
                               elem["input"]["config_tuna_id"])
         for solver in elem["applicable_solvers"]:
           try:
-            solver_id = solver_id_map_h[solver]
+            solver_id = self.solver_id_map[solver]
             obj = app_query.filter(
                 self.dbt.solver_app.solver == solver_id).first()  # pylint: disable=W0143
             if obj:
@@ -486,6 +477,184 @@ class FinClass(WorkerInterface):
           file for solvers that have been invalidated and are missing from your DB"
       )
       self.logger.info("Current invalid solvers: %s", solver_ids_invalid)
+
+    return True
+
+  def get_fdb_entry(self, session, solver):
+    """ Get FindDb entry from db """
+    fdb_entry = self.dbt.find_db_table()
+    fdb_entry.config = self.config.id
+    fdb_entry.solver = solver
+    fdb_entry.session = self.dbt.session.id
+    fdb_entry.opencl = False
+    fdb_entry.logger = self.logger
+    fdb_query = fdb_entry.get_query(session, self.dbt.find_db_table,
+                                    self.dbt.session.id)
+    obj = fdb_query.first()
+    return obj, fdb_entry
+
+  def update_fdb_entry(self, session, solver):
+    """ Add a new entry to fdb if there isnt one already """
+    obj, fdb_entry = self.get_fdb_entry(session, solver)
+    if obj:  # existing entry in db
+      # This can be removed if we implement the delete orphan cascade
+      fdb_entry = obj
+      session.query(
+          self.dbt.kernel_cache).filter(self.dbt.kernel_cache.kernel_group ==
+                                        fdb_entry.kernel_group).delete()
+    else:
+      # Insert the above entry
+      session.add(fdb_entry)
+    return fdb_entry
+
+  def compose_fdb_entry(self, session, fin_json, fdb_obj):
+    """Compose a FindDB table entry from fin_output"""
+    solver = self.solver_id_map[fdb_obj['solver_name']]
+    fdb_entry = self.update_fdb_entry(session, solver)
+    fdb_entry.fdb_key = fin_json['db_key']
+    fdb_entry.alg_lib = fdb_obj['algorithm']
+    fdb_entry.params = fdb_obj['params']
+    fdb_entry.workspace_sz = fdb_obj['workspace']
+    fdb_entry.valid = True
+
+    fdb_entry.kernel_time = -1
+    if 'time' in fdb_obj:
+      fdb_entry.kernel_time = fdb_obj['time']
+
+    fdb_entry, _ = self.get_fdb_entry(session, solver)
+    fdb_entry.kernel_group = fdb_entry.id
+
+    return fdb_entry
+
+  def compose_kernel_entry(self, session, fdb_obj, fdb_entry):
+    """Compose a new Kernel Cache entry from fin input"""
+    # Now we have the ID, lets add the binary cache objects
+    for kern_obj in fdb_obj['kernel_objects']:
+      kernel_obj = self.dbt.kernel_cache()
+      self.populate_kernels(kern_obj, kernel_obj)
+      kernel_obj.kernel_group = fdb_entry.kernel_group
+      session.add(kernel_obj)
+    return True
+
+  def populate_kernels(self, kern_obj, kernel_obj):
+    """populate kernel object"""
+    kernel_obj.kernel_name = kern_obj['kernel_file']
+    kernel_obj.kernel_args = kern_obj['comp_options']
+    kernel_obj.kernel_blob = bytes(kern_obj['blob'], 'utf-8')
+    kernel_obj.kernel_hash = kern_obj['md5_sum']
+    kernel_obj.uncompressed_size = kern_obj['uncompressed_size']
+    return kernel_obj
+
+  def update_fdb_w_kernels(self,
+                           session,
+                           fin_json,
+                           result_str='miopen_find_compile_result',
+                           check_str='find_compiled'):
+    """update find db + kernels from json results"""
+    status = []
+    if fin_json[result_str]:
+      for fdb_obj in fin_json[result_str]:
+        slv_stat = get_fin_slv_status(fdb_obj, check_str)
+        status.append(slv_stat)
+
+        if fdb_obj[check_str]:
+          #returned entry is added to the table
+          fdb_entry = self.compose_fdb_entry(session, fin_json, fdb_obj)
+          if fdb_obj['reason'] == 'Success':
+            self.compose_kernel_entry(session, fdb_obj, fdb_entry)
+            self.logger.info('Updating find Db(Build) for job_id=%s',
+                             self.job.id)
+          else:
+            # JD: add info about reason to the logs table
+            fdb_entry.valid = False
+        else:
+          self.logger.warning("Failed find_db compile, cfg_id: %s, obj: %s",
+                              fin_json['config_tuna_id'], fdb_obj)
+    else:
+      status = [{
+          'solver': 'all',
+          'success': False,
+          'result': 'Find Compile: No results'
+      }]
+
+    session.commit()
+
+    return status
+
+  def process_fdb_w_kernels(self,
+                            session,
+                            fin_json,
+                            result_str='miopen_find_compile_result',
+                            check_str='find_compiled'):
+    """initiate find db update"""
+
+    callback = self.update_fdb_w_kernels
+    status = session_retry(
+        session, callback,
+        lambda x: x(session, fin_json, result_str, check_str), self.logger)
+
+    if not status:
+      self.logger.warning('Fin: Unable to update Database')
+      status = [{
+          'solver': 'all',
+          'success': False,
+          'result': 'Fin: Unable to update Database'
+      }]
+
+    return status
+
+  def run_fin_cmd(self):
+    """Run a fin command after generating the JSON"""
+    fin_output = self.machine.make_temp_file()
+    cmd = []
+
+    env_str = " ".join(self.envmt)
+    cmd.append(env_str)
+    cmd.extend(
+        ['/opt/rocm/bin/fin', '-i',
+         self.get_fin_input(), '-o', fin_output])  # pylint: disable=no-member
+
+    for i in range(MAX_JOB_RETRIES):
+      ret_code, _, err = self.exec_docker_cmd(cmd)
+
+      if ret_code != 0:
+        self.logger.error('Error executing command: %s', ' '.join(cmd))
+        if err:
+          err_str = err.read()
+          self.logger.error('%s : %s', ret_code, err_str)
+          if "disk I/O error" in err_str:
+            self.logger.error('fin retry : %u', i)
+            sleep(random.randint(1, 10))
+          else:
+            break
+        else:
+          self.logger.error('err code : %s', ret_code)
+          break
+      else:
+        break
+
+    if ret_code != 0:
+      return None
+
+    # load the output json file and strip the env
+    fin_json = json.loads(self.machine.read_file(fin_output))[1:]
+    assert len(fin_json) == 1
+    # JD: if we implement multiple jobs per fin launch, this would be a loop
+    fin_json = fin_json[0]
+    return fin_json
+
+  def check_env(self):
+    """Checking that presumed rocm/miopen_v corresponds to the env rocm/miopen_v"""
+    env_rocm_v = self.get_rocm_v()
+    if self.dbt.session.rocm_v != env_rocm_v:
+      raise ValueError(
+          f'session rocm_v {self.dbt.session.rocm_v} does not match env rocm_v {env_rocm_v}'
+      )
+    env_miopen_v = self.get_miopen_v()
+    if self.dbt.session.miopen_v != env_miopen_v:
+      raise ValueError(
+          f'session rocm_v {self.dbt.session.rocm_v} does not match env rocm_v {env_rocm_v}'
+      )
 
     return True
